Added Navigation to sample app. Moved ComposableResources to its own file. Sample app now has screen with buttons to direct to screens.
